VOTE-1651:  update external links test (#321)
