--- testing/cypress/e2e/excludedLinks/excluded-links.cy.js
@@ -1,16 +1,0 @@
-/// <reference types="Cypress" />
-
-// this test will run out of the pipeline and will be a manual test each sprint 
-
-const english = require("../../fixtures/excluded-links-page.json");
-
-
-describe('check excluded links', () => {
-  english.forEach(url => {
-      it('validate page loads', () => {
-        cy.visit(url, {failOnStatusCode: false})
-        cy.get('body').children().its('length').should('be.gt', 0)
-        
-      })
-  })
-})
