--- testing/cypress/e2e/externalLinks/english-external-links-validator.cy.js
@@ -1,85 +1,0 @@
-/// <reference types="Cypress" />
-
-
-const allPages = require("../../fixtures/site-pages.json");
-
-const excludedlinks = [
-  'https://voterregistration.ct.gov/OLVR/welcome.do?ref=voteusa_en',
-  'https://voterregistration.ct.gov/OLVR/welcome.do',
-  // the above link will throw error code "read ECONNRESET" this will not pass through cypress test and has been checked manually but has not been added to `excluded-links.cy.js` test 
-  'https://voterservices.elections.maryland.gov/OnlineVoterRegistration/InstructionsStep1?ref=voteusa_en',
-  'https://www.sec.state.ma.us/ovr/?ref=voteusa_en',
-  'https://olvr.ohiosos.gov/?ref=voteusa_en',
-  'https://voterlookup.ohiosos.gov/voterlookup.aspx?ref=voteusa_en',
-  'https://olvr.hawaii.gov/?ref=voteusa_en',
-  'https://elections.hawaii.gov/voters/registration/?ref=voteusa_en',
-  'https://vote.sos.ri.gov/Home/RegistertoVote?ref=voteusa_en',
-  'https://vote.sos.ri.gov/Voter/RegisterToVote?ref=voteusa_en',
-  'https://vote.sos.ri.gov/Home/UpdateVoterRecord?ActiveFlag=0&?ref=voteusa_en',
-  'https://my.arizona.vote/WhereToVote.aspx?s=individual&?ref=voteusa_en',
-  'https://sosmt.gov/elections/vote/?ref=voteusa_en',
-  'https://www.vivote.gov/voters/register-to-vote/?ref=voteusa_en',
-  'https://www.vivote.gov/voters/voter-lookup/?ref=voteusa_en',
-  'https://www.elections.ny.gov/votingregister.html?ref=voteusa_en',
-  'https://voterlookup.elections.ny.gov/?ref=voteusa_en',
-  // todo: look into these links
-  'https://azsos.gov/elections/voting-election/register-vote-or-update-your-current-voter-information',
-  'https://my.arizona.vote/WhereToVote.aspx?s=individual',
-  'https://olvr.hawaii.gov/',
-  'https://www.elections.ny.gov/votingregister.html',
-  'https://olvr.ohiosos.gov/',
-  'https://vote.sos.ri.gov/Home/RegistertoVote',
-  'https://vivote.gov/voters/voter-lookup/',
-  'https://elections.hawaii.gov/voters/registration/',
-  'https://voterlookup.elections.ny.gov/',
-  'https://voterlookup.ohiosos.gov/voterlookup.aspx',
-  'https://vote.sos.ri.gov/Voter/RegisterToVote',
-  'https://vote.sos.ri.gov/Home/UpdateVoterRecord?ActiveFlag=0',
-  'https://vivote.gov/voters/register-to-vote/',
-  // sc links
-  'https://vrems.scvotes.sc.gov/Voter/Login?ref=voteusa_en',
-  'https://scvotes.gov/voters/register-to-vote/?ref=voteusa_en',
-  'https://info.scvotes.sc.gov/eng/ovr/start.aspx?ref=voteusa_en'
-
-];
-
-describe("External Link Validator Test", () => {
-  const singlePage =
-    Cypress.env("name") && Cypress.env("route")
-    ? [
-                {
-                  name: Cypress.env("name"),
-                  route: Cypress.env("route"),
-                },
-              ]
-      : null;
-  const pages = singlePage !== null ? singlePage : allPages;
-  pages.forEach((page) => {
-    it( 
-      `${page.name === "" ? "home" : page.name}`,
-      () =>
-        Cypress.env("retries") === true
-          ? {
-              retries: {
-                runMode: 2,
-              },
-            }
-          : {},
-      () => {
-        cy.visit({
-          url: page.route,
-        });
-        cy.get("main a[href^='https://']").each(link => {
-          if (excludedlinks.indexOf(link.prop('href')) == -1) {
-            cy.request({
-              url: link.prop('href'),
-              failOnStatusCode: false
-            }).then((response) => {
-              expect(response.status).to.eq(200)
-            })
-          }
-        })
-      }
-    );
-  });    
-});
