--- testing/cypress/fixtures/excluded-links-page.json
@@ -1,68 +1,0 @@
-[
-  "_comment", "English links below",
-
-  "https://voterservices.elections.maryland.gov/OnlineVoterRegistration/InstructionsStep1?ref=voteusa_en",
-  "https://www.sec.state.ma.us/ovr/?ref=voteusa_en",
-  "https://olvr.ohiosos.gov/?ref=voteusa_en",
-  "https://voterlookup.ohiosos.gov/voterlookup.aspx?ref=voteusa_en",
-  "https://olvr.hawaii.gov/?ref=voteusa_en",
-  "https://elections.hawaii.gov/voters/registration/?ref=voteusa_en",
-  "https://vote.sos.ri.gov/Home/RegistertoVote?ref=voteusa_en",
-  "https://vote.sos.ri.gov/Voter/RegisterToVote?ref=voteusa_en",
-  "https://vote.sos.ri.gov/Home/UpdateVoterRecord?ActiveFlag=0&?ref=voteusa_en",
-  "https://my.arizona.vote/WhereToVote.aspx?s=individual&?ref=voteusa_en",
-  "https://sosmt.gov/elections/vote/?ref=voteusa_en",
-  "https://www.vivote.gov/voters/register-to-vote/?ref=voteusa_en",
-  "https://www.vivote.gov/voters/voter-lookup/?ref=voteusa_en",
-  "https://www.elections.ny.gov/votingregister.html?ref=voteusa_en",
-  "https://voterlookup.elections.ny.gov/?ref=voteusa_en",
-  "https://azsos.gov/elections/voting-election/register-vote-or-update-your-current-voter-information",
-  "https://my.arizona.vote/WhereToVote.aspx?s=individual",
-  "https://olvr.hawaii.gov/",
-  "https://www.elections.ny.gov/votingregister.html",
-  "https://olvr.ohiosos.gov/",
-  "https://vote.sos.ri.gov/Home/RegistertoVote",
-  "https://vivote.gov/voters/voter-lookup/",
-  "https://elections.hawaii.gov/voters/registration/",
-  "https://voterlookup.elections.ny.gov/",
-  "https://voterlookup.ohiosos.gov/voterlookup.aspx",
-  "https://vote.sos.ri.gov/Voter/RegisterToVote",
-  "https://vote.sos.ri.gov/Home/UpdateVoterRecord?ActiveFlag=0",
-  "https://vrems.scvotes.sc.gov/Voter/Login?ref=voteusa_en",
-  "https://scvotes.gov/voters/register-to-vote/?ref=voteusa_en",
-  "https://info.scvotes.sc.gov/eng/ovr/start.aspx?ref=voteusa_en",
-  "https://vivote.gov/voters/register-to-vote/",
-
-  "_comment", "Spanish links below",
-
-  "https://voterservices.elections.maryland.gov/OnlineVoterRegistration/InstructionsStep1?ref=voteusa_es",
-  "https://www.sec.state.ma.us/ovr/?ref=voteusa_es",
-  "https://olvr.ohiosos.gov/?ref=voteusa_es",
-  "https://voterlookup.ohiosos.gov/voterlookup.aspx?ref=voteusa_es",
-  "https://olvr.hawaii.gov/?ref=voteusa_es",
-  "https://elections.hawaii.gov/voters/registration/?ref=voteusa_es",
-  "https://vote.sos.ri.gov/Home/RegistertoVote?ref=voteusa_es",
-  "https://vote.sos.ri.gov/VoterSpanish/RegisterToVote?ref=voteusa_es",
-  "https://vote.sos.ri.gov/HomeSpanish/UpdateVoterRecord?ActiveFlag=0&?ref=voteusa_es",
-  "https://vrems.scvotes.sc.gov/Voter/Login?ref=voteusa_es",
-  "https://scvotes.gov/voters/register-to-vote/?ref=voteusa_es",
-  "https://info.scvotes.sc.gov/eng/ovr/start.aspx?ref=voteusa_es",
-  "https://my.arizona.vote/WhereToVote.aspx?s=individual&?ref=voteusa_es",
-  "https://sosmt.gov/elections/vote/?ref=voteusa_es",
-  "https://www.vivote.gov/voters/register-to-vote/?ref=voteusa_es",
-  "https://www.vivote.gov/voters/voter-lookup/?ref=voteusa_es",
-  "https://www.elections.ny.gov/votingregister.html?ref=voteusa_es",
-  "https://voterlookup.elections.ny.gov/?ref=voteusa_es",
-  "https://azsos.gov/elections/voting-election/register-vote-or-update-your-current-voter-information-sp",
-  "https://olvr.hawaii.gov/",
-  "https://www.elections.ny.gov/votingregister.html",
-  "https://olvr.ohiosos.gov/",
-  "https://vote.sos.ri.gov/Home/RegistertoVote",
-  "https://vivote.gov/voters/voter-lookup/",
-  "https://my.arizona.vote/WhereToVote.aspx?s=individual",
-  "https://elections.hawaii.gov/voters/registration/",
-  "https://voterlookup.elections.ny.gov/",
-  "https://voterlookup.ohiosos.gov/voterlookup.aspx",
-  "https://vote.sos.ri.gov/VoterSpanish/RegisterToVote",
-  "https://vote.sos.ri.gov/HomeSpanish/UpdateVoterRecord?ActiveFlag=0"
-]
