+++ testing/cypress/support/e2e.js
@@ -11,3 +11,4 @@ return false;});
 import 'cypress-mochawesome-reporter/register';
 
 import "cypress-real-events";
+
